refactor: bring database insert to a single mapper
