+++ lib/src/features/menu/data/data_sources/savable_categories_data_source.dart
@@ -3,6 +3,7 @@ import 'package:drift/drift.dart';
 import '../../../../common/database/database.dart';
 import '../../models/dto/menu_category_dto.dart';
 import 'categories_data_source.dart';
+import '../../utils/category_database_mapper.dart';
 
 abstract interface class ISavableCategoriesDataSource
     implements ICategoriesDataSource {
