+++ lib/src/features/menu/models/dto/menu_item_dto.dart
@@ -1,4 +1,3 @@
-import 'package:drift/drift.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 import '../../../../common/database/database.dart';
@@ -25,7 +24,8 @@ class MenuItemDto {
   factory MenuItemDto.fromJson(Map<String, dynamic> json) =>
       _$MenuItemDtoFromJson(json);
 
-  static MenuItemDto fromDatabase(MenuItem item, MenuCategory category, List<MenuItemPrice> prices) {
+  static MenuItemDto fromDatabase(
+      MenuItem item, MenuCategory category, List<MenuItemPrice> prices) {
     return MenuItemDto(
       id: item.id,
       name: item.name,
@@ -35,20 +35,12 @@ class MenuItemDto {
         'slug': category.name,
       },
       imageUrl: item.imageUrl,
-      prices: prices.map((price) => {
-        'value':  price.value.toStringAsFixed(2),
-        'currency': price.currency,
-      }).toList(),
-    );
-  }
-
-  MenuItemsCompanion toDatabase() {
-    return MenuItemsCompanion.insert(
-      id: Value(id),
-      name: name,
-      description: Value(description),
-      imageUrl: Value(imageUrl),
-      categoryId: category['id'] as int,
+      prices: prices
+          .map((price) => {
+                'value': price.value.toStringAsFixed(2),
+                'currency': price.currency,
+              })
+          .toList(),
     );
   }
 }
